fix: consolidate messages for unpublished packages (#1761)
We currently send a message per unpublished version. However, a package
that has many versions unpublished will pollute the logs with many
messages. This PR consolidates those messages into one message that can
be collapsed so that the logs are much easier to parse.

No tests; this only affects logging

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*