fix: npmjs package canary can report undefined publish time (#1766)
It's possible for the latest version in
`https://registry.npmjs.org/construct-hub-probe/latest` to not be
included in the publish times that
`https://registry.npmjs.org/construct-hub-probe` reports. This is likely
because the package was just published.

This change protects from that scenario by failing the invocation. The
next invocation will likely pick up the publish time of the version
without our intervention.

Before this change, an undefined `publishedAt` property would result in
`publishedAt: "1970-01-01T00:00:00.000Z"` written into the state, which
is then used to erroneously determine the time between npm publish and
catalog availability.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>