feat: configurable alarm severities (#1768)
Followup on an internal team discussion:

### Backend Orchestration Failed

This alarm fires on a single execution failure -> setting it to a high
severity is too harsh because sporadic singular failures can happen from
time to time, and its not an indication of a severe problem.

Given we also have a high severity alarm for when execution failure rate
is above 75%, we want to set this alarm to be low severity.

### New version visibility SLA breached

This alarm fires when the package canary is not ingested in time,
according to the defined SLA. It is, presumably, an indication of a
major problem because the package is predictable and processable.
However, this alarm is also impacted by the npm replica, which, from to
time, experiences significant delays out of our control. We therefore
decided to set it to a low severity initally.

Since then, we have increased our SLA and the replica itself has also
undergone substantial changes. Seems like an appropriate timing to start
increasing the severity.

### Note

We could have also just changed the severities, without providing a way
to configure them. This however would be a breaking change. Construct
hub is not a stable package, so we would have gotten away with it, but I
decide to play it safe.
Configurable severities seems like a good feature to have anyway. 

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>