fix: DLQ alarm can fire spuriously
The DLQ alarms includes the `ApproximateMessagesNotVisible` metric.
That metric can behave unpredictably, sometimes approximating 1 even
if there are no messages in the DLQ at all.

Since there are no readers on a DLQ (usually), the number of messages
in flight will usually be 0. No need to include this in the alarm then.