fix: DLQ alarm can fire spuriously (#1779)
The DLQ alarms includes the `ApproximateMessagesNotVisible` metric. That
metric can behave unpredictably, sometimes approximating 1 even if there
are no messages in the DLQ at all.

Since there are no readers on a DLQ (usually), the number of messages in
flight will usually be 0. No need to include this in the alarm then.


----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>